requirements.txt
fastapi
openai
google-generativeai
instructor
firebase-admin
python-multipart
opencv-python
uvicorn 
gunicorn 
pillow